+++ lib/scout_apm/tracked_request.rb
@@ -43,13 +43,14 @@ module ScoutApm
 
     def initialize
       @layers = []
-      @call_counts = Hash.new { |h, k| h[k] = CallSet.new }
+      @call_set = Hash.new { |h, k| h[k] = CallSet.new }
       @annotations = {}
       @ignoring_children = false
       @context = Context.new
       @root_layer = nil
       @stackprof = nil
       @error = false
+      @mem_start = mem_usage
     end
 
     def start_layer(layer)
@@ -58,7 +59,6 @@ module ScoutApm
       end
 
       start_request(layer) unless @root_layer
-      update_call_counts!(layer)
       @layers[-1].add_child(layer) if @layers.any?
       @layers.push(layer)
     end
@@ -68,7 +68,11 @@ module ScoutApm
 
       layer = @layers.pop
       layer.record_stop_time!
+      layer.record_allocations!
 
+      # This must be called before checking if a backtrace should be collected as the call count influences our capture logic.
+      # We call `#update_call_counts in stop layer to ensure the layer has a final desc. Layer#desc is updated during the AR instrumentation flow.
+      update_call_counts!(layer)
       if capture_backtrace?(layer)
         layer.capture_backtrace!
       end
@@ -76,6 +80,18 @@ module ScoutApm
       if finalized?
         stop_request
       end
+    end
+
+    # Grab the currently running layer. Useful for adding additional data as we
+    # learn it. This is useful in ActiveRecord instruments, where we start the
+    # instrumentation early, and gradually learn more about the request that
+    # actually happened as we go (for instance, the # of records found, or the
+    # actual SQL generated).
+    # 
+    # Returns nil in the case there is no current layer. That would be normal
+    # for a completed TrackedRequest
+    def current_layer
+      @layers.last
     end
 
     BACKTRACE_BLACKLIST = ["Controller", "Job"]
@@ -93,7 +109,7 @@ module ScoutApm
       return true if layer.total_exclusive_time > BACKTRACE_THRESHOLD
 
       # Capture any layer that we've seen many times. Captures n+1 problems
-      return true if @call_counts[layer.name].capture_backtrace?
+      return true if @call_set[layer.name].capture_backtrace?
 
       # Don't capture otherwise
       false
@@ -101,7 +117,16 @@ module ScoutApm
 
     # Maintains a lookup Hash of call counts by layer name. Used to determine if we should capture a backtrace.
     def update_call_counts!(layer)
-      @call_counts[layer.name].update!(layer.desc)
+      @call_set[layer.name].update!(layer.desc)
+    end
+
+    # This may be in bytes or KB based on the OSX. We store this as-is here and only do conversion to MB in Layer Converters.
+    def mem_usage
+      ScoutApm::Instruments::Process::ProcessMemory.rss
+    end
+
+    def capture_mem_delta!
+      @mem_delta = mem_usage - @mem_start
     end
 
     ###################################
@@ -204,6 +229,9 @@ module ScoutApm
 
       slow_job = LayerConverters::SlowJobConverter.new(self).call
       ScoutApm::Agent.instance.store.track_slow_job!(slow_job)
+
+      allocation_metrics = LayerConverters::AllocationMetricConverter.new(self).call
+      ScoutApm::Agent.instance.store.track!(allocation_metrics)
     end
 
     # Have we already persisted this request?