+++ CHANGELOG.markdown
@@ -1,3 +1,18 @@
+# 2.0.0
+
+* Reporting object allocation metrics and mem delta with slow requests and jobs.
+
+# 1.5.2
+
+* Fix deploy webhook endpoint
+
+# 1.5.1
+
+* Collecting memory metrics on slow transactions
+* Collecting additional fields for slow transactions:
+  * hostname
+  * seconds_since_startup (larger memory increases and other other odd behavior more common when close to startup)
+
 # 1.5.0
 
 * Background Job instrumentation for Sidekiq and Sidekiq-backed ActiveJob